feat: move init generator to webpack-addons
