+++ lib/creator/index.js
@@ -1,8 +1,8 @@
 const yeoman = require('yeoman-environment');
 const Generator = require('yeoman-generator');
 const path = require('path');
-const defaultGenerator = require('./yeoman/webpack-generator');
-const WebpackAdapter = require('./yeoman/webpack-adapter');
+const defaultGenerator = require('webpack-addons').WebpackGenerator;
+const WebpackAdapter = require('./utils/webpack-adapter');
 const runTransform = require('./transformations/index');
 
 /*
