+++ lib/creator/utils/webpack-adapter.js
