--- lib/creator/yeoman/utils/module.js
@@ -1,12 +1,0 @@
-module.exports = () => {
-	return {
-		test: new RegExp(/\.js$/),
-		exclude: '/node_modules/',
-		loader: '\'babel-loader\'',
-		options: {
-			presets: [
-				'\'es2015\''
-			]
-		}
-	};
-};
