--- lib/creator/yeoman/utils/plugins.js
@@ -1,5 +1,0 @@
-module.exports = () => {
-	return [
-		'new UglifyJSPlugin()'
-	];
-};
