--- lib/creator/yeoman/utils/tooltip.js
@@ -1,49 +1,0 @@
-module.exports = {
-	uglify: () => {
-		return (`/*
- * We've enabled UglifyJSPlugin for you! This minifies your app
- * in order to load faster and run less javascript.
- *
- * https://github.com/webpack-contrib/uglifyjs-webpack-plugin
- *
- */`);
-	},
-	commonsChunk: () => {
-		return (`/*
- * We've enabled commonsChunkPlugin for you. This allows your app to
- * load faster and it splits the modules you provided as entries across
- * different bundles!
- *
- * https://webpack.js.org/plugins/commons-chunk-plugin/
- *
- */`);
-	},
-	cssPlugin: () => {
-		return(
-			`/*
- * We've enabled ExtractTextPlugin for you. This allows your app to
- * use css modules that will be moved into a separate CSS file instead of inside
- * one of your module entries!
- *
- * https://github.com/webpack-contrib/extract-text-webpack-plugin
- *
- */`);
-	},
-	postcss: () => {
-		return(
-			`/*
- * We've enabled Postcss, autoprefixer and precss for you. This allows your app
- * to lint  CSS, support variables and mixins, transpile future CSS syntax,
- * inline images, and more!
- *
- * To enable SASS or LESS, add the respective loaders to module.rules
- *
- * https://github.com/postcss/postcss
- *
- * https://github.com/postcss/autoprefixer
- *
- * https://github.com/jonathantneal/precss
- *
- */`);
-	}
-};
