--- lib/creator/yeoman/utils/validate.js
@@ -1,7 +1,0 @@
-module.exports = (value) => {
-	const pass = value.length;
-	if(pass) {
-		return true;
-	}
-	return 'Please specify an answer!';
-};
