+++ package.json
@@ -49,7 +49,7 @@
     "resolve-cwd": "^2.0.0",
     "supports-color": "^3.1.2",
     "webpack": "^2.5.1",
-    "webpack-addons": "^1.1.2",
+    "webpack-addons": "git://github.com/webpack-contrib/webpack-addons.git#yeoman-migration",
     "yargs": "^6.5.0",
     "yeoman-environment": "^1.6.6",
     "yeoman-generator": "git://github.com/ev1stensberg/generator.git#Feature-getArgument"
